refactor(rb-enrich): replace AnthropicEnricher with OpenAiCompatEnricher
Rename anthropic.rs → openai_compat.rs (git mv, history preserved).
AnthropicEnricher → OpenAiCompatEnricher using OpenAI chat-completions
protocol. POST to {base_url}/chat/completions with Bearer auth.

Activation now requires RB_ENRICH_BASE_URL + RB_ENRICH_MODEL (explicit
two-var opt-in); RB_ENRICH_API_KEY is optional (local servers like Ollama
need none). from_env() returns Ok(None) when either required var is absent.

No Debug derive, SecretString, key exposed only at header insertion, 30s
timeout, error_for_status, best-effort (engine falls back to heuristic).
Tests updated: wiremock against POST /chat/completions, Authorization:
Bearer header assertion, no-key path, key-leak assertion, env-var tests.
One #[ignore] real-API smoke test gated on RB_ENRICH_* vars.
